further updated for pathing. changed percision of many things to double that is, position and dimension introduced bug related to gridmanager where the top and right parts of an object in the gamearea will not be cliked, with probable further consequnces introduced AbstractPathObject. see docs. introduced Point. mostly and alias for Point2D.Double. used in percision change

+++ src/assault/display/Button.java
@@ -25,7 +25,7 @@ public abstract class Button extends Paintable implements MouseListener, Keyboar
     private Runnable action;
     private boolean enabled = true;
 
-    public Button(String name, int x, int y, int width, int height, char shortCut, BufferedImage icon, Runnable action) {
+    public Button(String name, double x, double y, double width, double height, char shortCut, BufferedImage icon, Runnable action) {
         super(x, y, width, height);
         this.name = name;
         setAction(action);
